requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
torch
transformers
tokenizers
scikit-learn
numpy
pandas
detoxify
kafka-python
sqlalchemy
psycopg2-binary
alembic
redis
streamlit
plotly
loguru
tqdm
pyyaml
orjson
passlib[bcrypt]
python-jose
httpx
requests
pytest
pytest-asyncio
black
isort
flake8